requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==1.9
Flask==1.0.2
Flask-Cors==3.0.4
Flask-SQLAlchemy==2.3.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.6.1
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
